platform/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
Pillow==11.1.0
numpy>=1.24,<2
opencv-python-headless>=4.8
pyyaml==6.0.2
mlflow==2.19.0
httpx==0.27.0
aiofiles==24.1.0
# For per-class validation metrics (CPU-only to keep image manageable)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu
torchvision==0.17.2+cpu
ultralytics==8.3.40
